Remove glob-reader dependency
This package was problematic for a couple of reasons.

First, it was specifying its dependencies as strict versions, rather
than with ranges. This makes upgrades problematic.

Second, it was forcibly stripping frontmatter from files. This impacts
line number references, which results in inaccurate feedback to users.

Third, it generally incldued a bunch of dependencies and other code I
simply didn't need. I've replaced it with a single tiny function,
without altering any of the tests. This was also important for the
"virtual filesystem" idea I've got planned for the future.